feat: add roles to the controller

+++ src/course-activities/course-activities.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   InternalServerErrorException,
   NotFoundException,
@@ -21,10 +22,16 @@ import {
 } from './course-activities.dto';
 import { isError } from '../utils/errors';
 import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
+import { CaslAbilityFactory } from '../casl/casl-ability.factory';
+import { request } from 'express';
+import { User } from '../db/entities/user.entity';
+import { E_ACTION } from '../casl/actions';
+import { filter } from 'lodash';
 
 @Controller('courses')
 export class CourseActivitiesController {
   constructor(
+    private readonly caslAbilityFactory: CaslAbilityFactory,
     private readonly courseActivitiesService: CourseActivitiesService,
   ) {}
 
@@ -40,10 +47,21 @@ export class CourseActivitiesController {
    * Find all activities for a course
    */
   @Get('/:abbr/activities')
+  @UseGuards(JwtAuthGuard)
   public async getAllForCourse(@Param('abbr') abbr: string) {
-    const foundCourse = this.courseActivitiesService.findByOptions({
-      [E_COURSE_ACTIVITY_ENTITY_KEYS.COURSE]: { abbr },
-    });
+    const rules = this.caslAbilityFactory.createForUser(request.user as User);
+
+    if (rules.cannot(E_ACTION.READ, CourseActivity))
+      throw new ForbiddenException(
+        "You don't have permission to read course activities",
+      );
+
+    const foundCourse = filter(
+      await this.courseActivitiesService.findByOptions({
+        [E_COURSE_ACTIVITY_ENTITY_KEYS.COURSE]: { abbr },
+      }),
+      (course_activity) => rules.can(E_ACTION.READ, course_activity),
+    );
 
     if (!foundCourse) throw new NotFoundException('Course not found');
 
@@ -56,6 +74,13 @@ export class CourseActivitiesController {
    */
   @Get('/activity/:id')
   public async getOne(@Param('id') id: string) {
+    const rules = this.caslAbilityFactory.createForUser(request.user as User);
+
+    if (rules.cannot(E_ACTION.READ, CourseActivity))
+      throw new ForbiddenException(
+        "You don't have permission to read course activities",
+      );
+
     return this.courseActivitiesService.findByOptions({
       [E_COURSE_ACTIVITY_ENTITY_KEYS.ID]: id,
     });
@@ -67,12 +92,30 @@ export class CourseActivitiesController {
    */
   @Post('/activity')
   public async create(@Body() createDto: CreateCourseActivitiesDto) {
-    return this.courseActivitiesService.create(createDto).catch((err) => {
-      if (isError(err, 'UNIQUE_CONSTRAINT'))
-        throw new NotFoundException('Course activity already exists');
-      else
-        throw new InternalServerErrorException("Can't create course activity");
-    });
+    const rules = this.caslAbilityFactory.createForUser(request.user as User);
+
+    if (rules.cannot(E_ACTION.CREATE, CourseActivity))
+      throw new ForbiddenException(
+        "You don't have permission to create course activity",
+      );
+
+    const createdCourseActivity = await this.courseActivitiesService
+      .create(createDto)
+      .catch((err) => {
+        if (isError(err, 'UNIQUE_CONSTRAINT'))
+          throw new NotFoundException('Course activity already exists');
+        else
+          throw new InternalServerErrorException(
+            "Can't create course activity",
+          );
+      });
+
+    if (!rules.can(E_ACTION.READ, createdCourseActivity))
+      throw new ForbiddenException(
+        "You don't have permission to read this course activity",
+      );
+
+    return createdCourseActivity;
   }
 
   /**
@@ -85,6 +128,13 @@ export class CourseActivitiesController {
     @Param('id') id: string,
     @Body() updateDto: UpdateCourseActivitiesDto,
   ) {
+    const rules = this.caslAbilityFactory.createForUser(request.user as User);
+
+    if (rules.cannot(E_ACTION.UPDATE, CourseActivity))
+      throw new ForbiddenException(
+        "You don't have permission to update course activity",
+      );
+
     const foundCourseActivity =
       await this.courseActivitiesService.findByOptions({
         [E_COURSE_ACTIVITY_ENTITY_KEYS.ID]: id,
@@ -93,11 +143,39 @@ export class CourseActivitiesController {
     if (!foundCourseActivity)
       throw new NotFoundException('Course activity not found');
 
-    return await this.courseActivitiesService.update(id, updateDto);
+    if (!rules.can(E_ACTION.UPDATE, foundCourseActivity[0]))
+      throw new ForbiddenException(
+        "You don't have permission to update this course activity",
+      );
+
+    const updatedCourseActivity = await this.courseActivitiesService
+      .update(id, updateDto)
+      .catch((err) => {
+        if (isError(err, 'UNIQUE_CONSTRAINT'))
+          throw new NotFoundException('Course activity already exists');
+        else
+          throw new InternalServerErrorException(
+            "Can't update course activity",
+          );
+      });
+
+    if (!rules.can(E_ACTION.READ, updatedCourseActivity))
+      throw new ForbiddenException(
+        "You don't have permission to read this course activity",
+      );
+
+    return updatedCourseActivity;
   }
 
   @Delete('/activity/:id')
   public async delete(@Param('id') id: string) {
+    const rules = this.caslAbilityFactory.createForUser(request.user as User);
+
+    if (rules.cannot(E_ACTION.DELETE, CourseActivity))
+      throw new ForbiddenException(
+        "You don't have permission to delete course activity",
+      );
+
     const foundCourseActivity =
       await this.courseActivitiesService.findByOptions({
         [E_COURSE_ACTIVITY_ENTITY_KEYS.ID]: id,
@@ -105,6 +183,11 @@ export class CourseActivitiesController {
 
     if (!foundCourseActivity)
       throw new NotFoundException('Course activity not found');
+
+    if (!rules.can(E_ACTION.DELETE, foundCourseActivity[0]))
+      throw new ForbiddenException(
+        "You don't have permission to delete this course activity",
+      );
 
     await this.courseActivitiesService.delete(id);
   }
